LPS-85088 Expand UploadPortletRequestTest into individual test classes
