+++ modules/apps/static/portal-osgi-web/portal-osgi-web-portlet-container-test/src/testIntegration/java/com/liferay/portal/osgi/web/portlet/container/upload/portlet/request/test/UploadPortletRequestWhenGettingFileAsStreamTest.java
@@ -1,0 +1,143 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+package com.liferay.portal.osgi.web.portlet.container.upload.portlet.request.test;
+
+import com.liferay.arquillian.extension.junit.bridge.junit.Arquillian;
+import com.liferay.portal.kernel.upload.FileItem;
+import com.liferay.portal.kernel.upload.UploadPortletRequest;
+import com.liferay.portal.test.rule.LiferayIntegrationTestRule;
+import com.liferay.portal.upload.LiferayServletRequest;
+import com.liferay.portal.upload.UploadPortletRequestImpl;
+import com.liferay.portal.upload.UploadServletRequestImpl;
+import com.liferay.portal.util.test.PortletContainerTestUtil;
+
+import java.io.InputStream;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+
+import javax.servlet.http.HttpServletRequest;
+import org.junit.Assert;
+import org.junit.Before;
+import org.junit.ClassRule;
+import org.junit.Rule;
+import org.junit.Test;
+import org.junit.rules.TestRule;
+import org.junit.runner.RunWith;
+
+/**
+ * @author Manuel de la Peña
+ */
+@RunWith(Arquillian.class)
+public class UploadPortletRequestWhenGettingFileAsStreamTest {
+
+	@ClassRule
+	@Rule
+	public static final TestRule aggregateTestRule =
+		new LiferayIntegrationTestRule();
+
+	@Before
+	public void setUp() throws Exception {
+		_setUp();
+	}
+
+	@Test
+	public void shouldReturnAStream() throws Exception {
+		Map<String, FileItem[]> fileParameters =
+			PortletContainerTestUtil.getFileParameters(
+				1, _portletNamespace, _BYTES);
+
+		LiferayServletRequest liferayServletRequest =
+			PortletContainerTestUtil.getMultipartRequest(
+				_portletNamespace, _BYTES);
+
+		UploadPortletRequest uploadPortletRequest =
+			new UploadPortletRequestImpl(
+				new UploadServletRequestImpl(
+					(HttpServletRequest)liferayServletRequest.getRequest(),
+					fileParameters, new HashMap<String, List<String>>()),
+				null, _portletNamespace);
+
+		Map<String, FileItem[]> multipartParametersMap =
+			uploadPortletRequest.getMultipartParameterMap();
+
+		Assert.assertEquals(
+			multipartParametersMap.toString(), 1,
+			multipartParametersMap.size());
+
+		for (Map.Entry<String, FileItem[]> entry :
+			multipartParametersMap.entrySet()) {
+
+			String key = entry.getKey();
+
+			InputStream inputStream = uploadPortletRequest.getFileAsStream(
+				key);
+
+			Assert.assertNotNull(inputStream);
+
+			inputStream = uploadPortletRequest.getFileAsStream(key, true);
+
+			Assert.assertNotNull(inputStream);
+		}
+	}
+
+	@Test
+	public void shouldReturnNullIfFileParametersAreEmpty()
+		throws Exception {
+
+		LiferayServletRequest liferayServletRequest =
+			PortletContainerTestUtil.getMultipartRequest(
+				_portletNamespace, _BYTES);
+
+		UploadPortletRequest uploadPortletRequest =
+			new UploadPortletRequestImpl(
+				new UploadServletRequestImpl(
+					(HttpServletRequest)liferayServletRequest.getRequest(),
+					new HashMap<String, FileItem[]>(),
+					new HashMap<String, List<String>>()),
+				null, _portletNamespace);
+
+		Assert.assertNull(
+			uploadPortletRequest.getFileAsStream("irrelevantName"));
+		Assert.assertNull(
+			uploadPortletRequest.getFileAsStream("irrelevantName", true));
+	}
+
+	@Test
+	public void shouldReturnNullIfNameIsNotAFileParameter()
+		throws Exception {
+
+		Map<String, FileItem[]> fileParameters =
+			PortletContainerTestUtil.getFileParameters(
+				1, _portletNamespace, _BYTES);
+
+		LiferayServletRequest liferayServletRequest =
+			PortletContainerTestUtil.getMultipartRequest(
+				_portletNamespace, _BYTES);
+
+		UploadPortletRequest uploadPortletRequest =
+			new UploadPortletRequestImpl(
+				new UploadServletRequestImpl(
+					(HttpServletRequest)liferayServletRequest.getRequest(),
+					fileParameters, new HashMap<String, List<String>>()),
+				null, _portletNamespace);
+
+		Assert.assertNull(
+			uploadPortletRequest.getFileAsStream("nonexistentFile"));
+		Assert.assertNull(
+			uploadPortletRequest.getFileAsStream("nonexistentFile", true));
+	}
+
+}
